% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dedup_history.R
\name{create_dedup_history}
\alias{create_dedup_history}
\title{Creates an excel file to store the deduplication history.}
\usage{
create_dedup_history(directory)
}
\arguments{
\item{directory}{Choose the directory in which
the references deduplication history will be saved.}
}
\value{
A list with elements:
\describe{
\item{history_dedup}{A Workbook object (from \pkg{openxlsx}).}
\item{hist_dedup_path}{Character. Path to the created .xlsx file.}
}
}
\description{
Creates an excel file to store the deduplication history.
}
\keyword{internal}
