% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_short_path.R
\name{get_short_path}
\alias{get_short_path}
\title{Transforms a long computer path into a shorter.}
\usage{
get_short_path(path)
}
\arguments{
\item{path}{Path to the document.}
}
\value{
A shorter path.
}
\description{
Transforms a long computer path into a shorter.
}
\keyword{internal}
