% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_scheduled_task.R
\name{remove_scheduled_task}
\alias{remove_scheduled_task}
\title{Removes a scheduled task.}
\usage{
remove_scheduled_task(task_id, dry_run = FALSE)
}
\arguments{
\item{task_id}{Name/ID of the scheduled task (Windows Task Scheduler or Cron).}

\item{dry_run}{Simulation run option.}
}
\value{
\code{NULL} (invisibly). Called for its side effects: removes a scheduled task saved using the function 'auto_LitFetchR_setup'.
}
\description{
Removes a scheduled task using the "task_id"
from Task Scheduler (Windows) or Cron (Mac/Linux).
}
\examples{
# This is a "dry run" example.
# No task will actually be removed, it only shows how the function should react.
remove_scheduled_task("fish_vibrio",
                      dry_run = TRUE
                      )

}
