% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_LitFetchR_setup.R
\name{auto_LitFetchR_setup}
\alias{auto_LitFetchR_setup}
\title{creates a read-only code file and a task to run the code automatically.}
\usage{
auto_LitFetchR_setup(
  reviewID = "reviewID",
  when = "DAILY",
  time = "08:00",
  WOS = TRUE,
  SCP = TRUE,
  PMD = TRUE
)
}
\arguments{
\item{reviewID}{name of the review}

\item{when}{frequency of the fetching, i.e. DAILY, WEEKLY, MONTHLY}

\item{time}{time of the fetching, i.e. HH:MM 24-hour clock format}

\item{WOS}{choose to search on Web of Science}

\item{SCP}{choose to search on Scopus}

\item{PMD}{choose to search on PubMed}
}
\value{
create a Rcode file and a task in Task Scheduler (Windows), or in Cron (Mac/Linux)
}
\description{
creates a read-only code file and a task to run the code automatically.
}
