% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_LitFetchR_setup.R
\name{auto_LitFetchR_setup}
\alias{auto_LitFetchR_setup}
\title{Automating the retrieval of references based on a saved search string}
\usage{
auto_LitFetchR_setup(
  task_ID = "task_ID",
  when = "DAILY",
  time = "08:00",
  WOS = TRUE,
  SCP = TRUE,
  PMD = TRUE,
  dedup = FALSE,
  open_file = FALSE
)
}
\arguments{
\item{task_ID}{Name of the automated task (e.g. one keyword describing your review)}

\item{when}{Frequency of the fetching, i.e. DAILY, WEEKLY, MONTHLY}

\item{time}{Time of the fetching, i.e. HH:MM 24-hour clock format}

\item{WOS}{Choose to search on Web of Science (TRUE or FALSE)}

\item{SCP}{Choose to search on Scopus (TRUE or FALSE)}

\item{PMD}{Choose to search on PubMed (TRUE or FALSE)}

\item{dedup}{Choose to deduplicate or not the references (TRUE or FALSE)}

\item{open_file}{choose to automatically open the CSV file after reference retrieval}
}
\value{
Create a Rscript file (READ ONLY) and a task in Task Scheduler (Windows), or in Cron (Mac/Linux)
}
\description{
\code{auto_LitFetchR_setup} creates a read-only Rscript and a task to run the code automatically
at a specified frequency and time, on up to three platforms (e.g. \href{https://clarivate.com/academia-government/scientific-and-academic-research/research-discovery-and-referencing/web-of-science/}{Web of Science},
\href{https://www.elsevier.com/en-au/products/scopus}{Scopus} and \href{https://pubmed.ncbi.nlm.nih.gov/}{PubMed})
}
\examples{
\dontrun{

#Example of what you should see:
auto_LitFetchR_setup(task_ID = "fish_vibrio",
                       when = "WEEKLY",
                       time = "14:00",
                       WOS = TRUE,
                       SCP = TRUE,
                       PMD = TRUE)
Task scheduled!
}

}
