% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dedup_history.R
\name{create_dedup_history}
\alias{create_dedup_history}
\title{creates an excel file to store the deduplication history}
\usage{
create_dedup_history()
}
\value{
create an excel file
}
\description{
creates an excel file to store the deduplication history
}
