% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dedup_history.R
\name{create_dedup_history}
\alias{create_dedup_history}
\title{Creates an excel file to store the deduplication history.}
\usage{
create_dedup_history()
}
\value{
The excel file 'history_dedup.xlsx'.
}
\description{
Creates an excel file to store the deduplication history.
}
\keyword{internal}
