% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id_history.R
\name{create_id_history}
\alias{create_id_history}
\title{creates an excel file to store the references' identification retrieved at each search}
\usage{
create_id_history()
}
\value{
create an excel file
}
\description{
creates an excel file to store the references' identification retrieved at each search
}
