% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_save_search.R
\name{create_save_search}
\alias{create_save_search}
\title{create and save search string}
\usage{
create_save_search()
}
\value{
one or more search string(s) in the file "search_list.txt" and the history of all searches in "history_search.xlsx"
}
\description{
\code{create_save_search}  is an interactive function that
ask the user to enter a search string and provide
the number of results from 3 platforms:
Web of Science, Scopus and PubMed. You can then save one
or more search strings to retrieve the references later.
}
\examples{
\dontrun{

# Example of what you should see building the search: fish AND "vibrio harveyi" AND diagnostic
> create_save_search()
History had been created.
Enter your search string (or 'summary' or 'exit'): fish
[1] "fish"
[1] "Web of Science: 1793296 results"
[1] "Scopus: 718644 results"
[1] "PubMed: 384742 results"
Enter your search string (or 'summary' or 'exit'): fish AND "vibrio harveyi"
[1] "fish AND \"vibrio harveyi\""
[1] "Web of Science: 2084 results"
[1] "Scopus: 1080 results"
[1] "PubMed: 727 results"
Enter your search string (or 'summary' or 'exit'): fish AND "vibrio harveyi" AND diagnostic
[1] "fish AND \"vibrio harveyi\" AND diagnostic"
[1] "Web of Science: 126 results"
[1] "Scopus: 22 results"
[1] "PubMed: 106 results"
Enter your search string (or 'summary' or 'exit'): summary
                               Search_Term Results_WOS Results_SCP Results_PMD
1                                     fish     1793296      718644      384742
2                fish AND "vibrio harveyi"        2084        1080         727
3 fish AND "vibrio harveyi" AND diagnostic         126          22         106
Select the index number for the search string to use in automated retrieval: 3

Selected search string: fish AND "vibrio harveyi" AND diagnostic
Do you want to save the search string for future use? (yes/no): yes
Enter a name for the search identification: fish_vibrio
Search string saved successfully.
Enter your search string (or 'summary' or 'exit'): exit
Exiting search tool.
}

}
