% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_save_search.R
\name{create_save_search}
\alias{create_save_search}
\title{Create and save search string(s).}
\usage{
create_save_search(dry_run = FALSE)
}
\arguments{
\item{dry_run}{Simulation run option.}
}
\value{
One or more search string(s) in the file "search_list.txt" and the history of all searches in "history_search.xlsx".
}
\description{
An interactive function that
ask the user to enter a search string and provide
the number of results from 3 platforms:
Web of Science, Scopus and PubMed. You can then save one
or more search strings to retrieve the references later.
}
\examples{
# This is a "dry run" example.
# No search will be created and no database will be accessed.
# It only shows how the function should react.
create_save_search(dry_run = TRUE)

}
