% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_search_history.R
\name{create_search_history}
\alias{create_search_history}
\title{creates an excel file to store the search history}
\usage{
create_search_history()
}
\value{
create an excel file
}
\description{
creates an excel file to store the search history
}
