% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_search_history.R
\name{create_search_history}
\alias{create_search_history}
\title{Creates an excel file to store the history of searches
made using \code{create_save_search()}}
\usage{
create_search_history()
}
\value{
"history_search.xlsx" and or a new sheet for the new search(es)
}
\description{
Creates an excel file to store the history of searches
made using \code{create_save_search()}
}
