% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{dedup_refs}
\alias{dedup_refs}
\title{deduplicate the references from three dataframes}
\usage{
dedup_refs(df1, df2, df3)
}
\arguments{
\item{df1}{dataframe 1}

\item{df2}{dataframe 2}

\item{df3}{dataframe 3}
}
\value{
A CSV file containing all the new references deduplicated
}
\description{
deduplicate the references from three dataframes
}
