% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{dedup_refs}
\alias{dedup_refs}
\title{Deduplicates the references from up to three dataframes.}
\usage{
dedup_refs(
  df1 = NULL,
  df2 = NULL,
  df3 = NULL,
  open_file = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{df1}{Dataframe 1 (can be NULL)}

\item{df2}{Dataframe 2 (can be NULL)}

\item{df3}{Dataframe 3 (can be NULL)}

\item{open_file}{Automatically opens the CSV file after reference retrieval.}

\item{dry_run}{Simulation run option.}
}
\value{
A CSV file containing all the new references deduplicated
and the history of the deduplication in an excel file.
}
\description{
Deduplicates the references from up to three dataframes.
}
\examples{
# This is a "dry run" example.
# No deduplication will happen.
# It only shows how the function should react.
dedup_refs(df_vibrio_wos,
           df_vibrio_scp,
           df_vibrio_pmd,
           dry_run = TRUE
           )

}
