% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{dedup_refs}
\alias{dedup_refs}
\title{deduplicate the references from three dataframes}
\usage{
dedup_refs(df1 = NULL, df2 = NULL, df3 = NULL, open_file = FALSE)
}
\arguments{
\item{df1}{dataframe 1, can be null}

\item{df2}{dataframe 2, can be null}

\item{df3}{dataframe 3, can be null}

\item{open_file}{choose to automatically open the CSV file after reference deduplication}
}
\value{
A CSV file containing all the new references deduplicated and the history of the deduplication.
}
\description{
deduplicate the references from three dataframes
}
\examples{
\dontrun{

#Example of what you should see:
> dedup_refs(df_vibrio_wos, df_vibrio_scp, df_vibrio_pmd)
Warning: The following columns are missing: pages, number, record_id, isbn
formatting data...
identifying potential duplicates...
identified duplicates!
flagging potential pairs for manual dedup...
Joining with `by = join_by(duplicate_id.x, duplicate_id.y)`
254 citations loaded...
14 duplicate citations removed...
240 unique citations remaining!
Deduplication script has been executed, concatenated deduplicated references had been exported.
Warning message:
In add_missing_cols(raw_citations) :
  Search contains missing values for the record_id column.
  A record_id will be created using row numbers
}

}
