% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pmd_list.R
\name{extract_pmd_list}
\alias{extract_pmd_list}
\title{Extracts the metadata from the new references found on PubMed
based on the search string(s) saved in "search_list.txt".}
\usage{
extract_pmd_list(search_list_path)
}
\arguments{
\item{search_list_path}{Path to "search_list.txt".}
}
\value{
A dataframe containing the metadata from the new references found on PubMed.
}
\description{
Extracts the metadata from the new references found on PubMed
based on the search string(s) saved in "search_list.txt".
}
\keyword{internal}
