% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_scp_list.R
\name{extract_scp_list}
\alias{extract_scp_list}
\title{extract the metadata from the new references from Scopus
based on the search strings found in search_list.txt}
\usage{
extract_scp_list(search_list_path)
}
\arguments{
\item{search_list_path}{path to search_list}
}
\value{
A dataframe containing the metadata from the new references on Scopus
}
\description{
extract the metadata from the new references from Scopus
based on the search strings found in search_list.txt
}
