% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wos_list.R
\name{extract_wos_list}
\alias{extract_wos_list}
\title{extract the metadata from the new references from Web of Science
based on the search strings found in search_list.txt}
\usage{
extract_wos_list(search_list_path)
}
\arguments{
\item{search_list_path}{path to search_list}
}
\value{
A dataframe containing the metadata from the new references on Web of Science
}
\description{
extract the metadata from the new references from Web of Science
based on the search strings found in search_list.txt
}
\keyword{internal}
