% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_short_path.R
\name{get_short_path}
\alias{get_short_path}
\title{transforms a long computer path into a shorter}
\usage{
get_short_path(path)
}
\arguments{
\item{path}{path to the document}
}
\value{
a shorter path
}
\description{
transforms a long computer path into a shorter
}
\keyword{internal}
