% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_fetch.R
\name{manual_fetch}
\alias{manual_fetch}
\title{manual literature fetch}
\usage{
manual_fetch(WOS = TRUE, SCP = TRUE, PMD = TRUE)
}
\arguments{
\item{WOS}{choose to search on Web of Science}

\item{SCP}{choose to search on Scopus}

\item{PMD}{choose to search on PubMed}
}
\value{
create a CSV file with the literature metadata
}
\description{
manual literature fetch
}
