% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_fetch.R
\name{manual_fetch}
\alias{manual_fetch}
\title{manual literature fetch}
\usage{
manual_fetch(WOS = TRUE, SCP = TRUE, PMD = TRUE)
}
\arguments{
\item{WOS}{choose to search on Web of Science (TRUE or FALSE)}

\item{SCP}{choose to search on Scopus (TRUE or FALSE)}

\item{PMD}{choose to search on PubMed (TRUE or FALSE)}
}
\value{
create a CSV file with the literature metadata, a history file of the references retreived and a history file of the deduplication
}
\description{
manual literature fetch
}
\examples{
\dontrun{

#Example of what you should see:
> manual_fetch(WOS = TRUE, SCP = TRUE, PMD = TRUE)
[1] 126
Finished batch number 1
Finished batch number 2
[1] "10.1016/j.aaf.2023.11.002 1 / 126"
[1] "10.3390/fishes10090439 2 / 126"
[FOR THE PURPOSE OF THIS EXAMPLE WE ARE NOT SHOWING EACH LINE OF THE CONSOLE]
[1] "NA 125 / 126"
[1] "NA 126 / 126"
[1] 22
Finished batch number 1
File already exists
[1] "10.1007/s12602-023-10207-x 1 / 22"
[1] "10.1016/j.fsi.2025.110189 2 / 22"
[FOR THE PURPOSE OF THIS EXAMPLE WE ARE NOT SHOWING EACH LINE OF THE CONSOLE]
[1] "10.1111/j.1472-765X.2010.02894.x 21 / 22"
[1] "NA 22 / 22"
[1] 106
Finished batch 1 for fish AND "vibrio harveyi" AND diagnostic
File already exists
[1] "10.1016/j.fsi.2025.110503 1 / 106"
[1] "10.1016/j.fsi.2025.110501 2 / 106"
[FOR THE PURPOSE OF THIS EXAMPLE WE ARE NOT SHOWING EACH LINE OF THE CONSOLE][1] "NA 105 / 106"
[1] "NA 106 / 106"
Warning: The following columns are missing: pages, number, record_id, isbn
formatting data...
identifying potential duplicates...
identified duplicates!
flagging potential pairs for manual dedup...
Joining with `by = join_by(duplicate_id.x, duplicate_id.y)`
254 citations loaded...
14 duplicate citations removed...
240 unique citations remaining!
Deduplication script has been executed, concatenated deduplicated references had been exported.
Warning message:
In add_missing_cols(raw_citations) :
  Search contains missing values for the record_id column. A record_id will be created using row numbers
}

}
