% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_scheduled_task.R
\name{remove_scheduled_task}
\alias{remove_scheduled_task}
\title{Remove a scheduled task by name from Task Scheduler (Windows) or Cron (Mac/Linux)}
\usage{
remove_scheduled_task(taskname)
}
\arguments{
\item{taskname}{Name/ID of the scheduled task (Windows Task Scheduler or cronR id)}
}
\description{
Remove a scheduled task by name from Task Scheduler (Windows) or Cron (Mac/Linux)
}
\examples{
\dontrun{

#Example of what you should see:
> remove_scheduled_task("fish_vibrio")
SUCCESS: The scheduled task "fish_vibrio" was successfully deleted.
Windows task 'fish_vibrio' removed (or did not exist).
}

}
